requirements.txt
requests
boto3
pycryptodome
